GH-9 renamed transaction component to Transactions
